git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@12935 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ src/com/l7tech/console/panels/OversizedTextDialog.java
@@ -87,8 +87,8 @@ public class OversizedTextDialog extends JDialog {
         Utilities.enableGrayOnDisabled(nestingDepthField);
         Utilities.enableGrayOnDisabled(payloadCountField);
 
-        validator.constrainTextFieldToNumberRange("text length", textLengthField, 0, Long.MAX_VALUE);
-        validator.constrainTextFieldToNumberRange("attribute length", attrLengthField, 0, Long.MAX_VALUE);
+        validator.constrainTextFieldToNumberRange("text length", textLengthField, 0, Integer.MAX_VALUE);
+        validator.constrainTextFieldToNumberRange("attribute length", attrLengthField, 0, Integer.MAX_VALUE);
         validator.constrainTextFieldToNumberRange("maximum payload count", payloadCountField, 0, Integer.MAX_VALUE);
         validator.constrainTextFieldToNumberRange("maximum nesting depth", nestingDepthField,
                                                   OversizedTextAssertion.MIN_NESTING_LIMIT,
